feat(6711): extract cli scripted driver root modules
